requirements.txt
asgiref==3.8.1
Django==5.1.5
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
PyJWT==2.10.1
python-dotenv==1.0.0
sqlparse==0.5.3
tzdata==2025.1
